+++ geocoding/src/types/GeoCodingFieldConfig.ts
@@ -2,7 +2,7 @@ import type { JSONField, PointField } from 'payload'
 
 /** Configuration for the geocoding fields. */
 export type GeoCodingFieldConfig = {
-  geoDataFieldOverride?: {
+  locationMetaOverride?: {
     access?: JSONField['access']
     admin?: JSONField['admin']
     label?: string